backend: Move LoanList system to common
